v2.17.17 chore(trip): remove 6 unreachable sheets + AccountPage avatar fix
User /qa 找到 critical issue:`/trip/:id` 永遠 redirect 到 TripsListPage
embedded mode,EmbeddedActionMenu 只暴露 5 項 (共編 + 列印 + 4 download),
其餘 6 個 sheet (AI 建議 / 今日路線 / 航班 / 出發清單 / 雨天備案 / 緊急)
完全沒 UI 入口。User 拍板「移除」。

Removed (Critical F1):
- OverflowMenu component (12 OVERFLOW_ITEMS dead)
- SuggestionSheet / FlightSheet / TodayRouteSheet / DocCard component
- TripSheetContent dispatcher (9 dead case branches + ACTION_MENU_GRID
  9-button + ACTION_MENU_EXPORTS 5-row)
- TripPage standalone TitleBar block 整段 (5 buttons + OverflowMenu)
- handlePanelItem / handleTripChange / sheetTrips / colorMode 等不再用 state
- ?sheet= URL allowlist 從 SHEET_TITLES 縮成 hardcode === 'collab'
- 4 dead test files

Fixed (High F2):
- AccountPage.tsx:236 avatar 用 displayName.charAt(0) 對齊 DesktopSidebar
  (原 email.charAt(0) 造成 hero 顯「L」 / sidebar 顯「R」 不一致)

Moved:
- DocEntry type 從 deleted DocCard 搬到 types/trip.ts (export flow 還用)

Stats: -14 files / -1500+ lines / precache 2189 → 2164 KiB / 1316 → 1258 tests
